+++ Sources/Indice.Identity/Client/Services/AuthorizationService.swift
@@ -7,7 +7,6 @@
 
 import Foundation
 import SwiftyJSON
-import IndiceNetworkClient
 
 /** Protocol containing any login method available on the Identity server. */
 public protocol AuthorizationService: AnyObject {
@@ -177,9 +176,6 @@ internal class AuthorizationServiceImpl: AuthorizationService {
             .with(deviceIds: thisDeviceRepository.ids)
         
         tokenStorage.parse(try await authRepository.authorize(grant: final))
-        
-        // try await refreshUserInfo()
-        // try await refreshDevices()
     }
     
     func tokenFor(authorizationDetails details: JSON, withGrant grant: OAuth2Grant) async throws -> TokenResponse {
@@ -193,7 +189,7 @@ internal class AuthorizationServiceImpl: AuthorizationService {
     
     public func refreshTokens() async throws {
         guard let refresh = tokenStorage.refreshToken else {
-            throw APIError.Unauthenticated
+            throw errorOfType(.authorization(error: .refreshTokenMissing))
         }
         
         try await login(withGrant: .refreshToken(with: refresh.value))
@@ -228,13 +224,17 @@ internal class AuthorizationServiceImpl: AuthorizationService {
      */
     func authorizationUrl(withPkce pkce: PKCE, andPrompt prompt: String) throws -> URL {
         guard var url = URL(string: configuration.authorizationEndpoint) else {
-            throw IdentityClient.Errors.AuthUrl
+            throw errorOfType(.url(malformedUrl: configuration.authorizationEndpoint))
+        }
+        
+        guard let redirectUri = client.urls?.authorization else {
+            throw errorOfType(.url(malformedUrl: configuration.authorizationEndpoint))
         }
         
         let queryParams = ["client_id"      : client.id,
                            "client_secret"  : client.secret,
                            "scope"          : client.userScope.value,
-                           "redirect_uri"   : client.urls.authorization,
+                           "redirect_uri"   : redirectUri,
                            "response_type"  : configuration.authCodeResponseType,
                            "response_mode"  : configuration.authCodeResponseMode,
                            "prompt"         : prompt,
@@ -261,12 +261,16 @@ internal class AuthorizationServiceImpl: AuthorizationService {
     
     func endSessionUrl() throws -> URL {
         guard var url = URL(string: configuration.logoutEndpoint) else {
-            throw IdentityClient.Errors.AuthUrl
+            throw errorOfType(.url(malformedUrl: configuration.logoutEndpoint))
+        }
+        
+        guard let postLogout = client.urls?.postLogout else {
+            throw errorOfType(.url(malformedUrl: configuration.logoutEndpoint))
         }
         
         let queryParams: [URLQueryItem] = [
             .init(name: "id_token_hint",            value: tokenStorage.idToken),
-            .init(name: "post_logout_redirect_uri", value: client.urls.postLogout)]
+            .init(name: "post_logout_redirect_uri", value: postLogout)]
         
         if #available(iOS 16.0, *) {
             url.append(queryItems: queryParams)
@@ -286,9 +290,9 @@ private extension AuthorizationServiceImpl {
         case .value(let pair): return pair
         case .error(let code):
             if code == errSecUserCanceled {
-                throw IdentityClient.Errors.UserCancel
+                throw errorOfType(.biometric(error: .userCanceled))
             } else {
-                throw IdentityClient.Errors.SecKeys
+                throw errorOfType(.biometric(error: .dataMissing))
             }
         }
     }